Create release manually
Rather than create a release for every push to main, create a release only when a tag matches a semver regex (^v[0-9]+.[0-9]+.[0-9]+$). This allows v1.2.3, but not v1.2.3-prerelease+3 to trigger a new release. If the release is already created, which might be the common workflow, then build artifacts will be attached to the existing release with that tag.